Rename: change directory controller.js
